+++ src/query/data/SubQuery.cpp
@@ -1,5 +1,6 @@
 #include "SubQuery.h"
 
+#include <cstdint>
 #include <memory>
 #include <numeric>
 #include <string>
@@ -34,12 +35,12 @@ QueryResult::Ptr SubQuery::execute() {
     if (condInit.second) {
       for (auto it = table.begin(); it != table.end(); ++it) {
         if (evalCondition(*it)) {
-          long long value = (*it)[srcId[0]];
+          int64_t value = (*it)[srcId[0]];
           // subtract the sum of remaining sources using std::accumulate
-          // use long long to avoid overflow during accumulation
-          long long sub_sum = std::accumulate(
+          // use int64_t to avoid overflow during accumulation
+          int64_t sub_sum = std::accumulate(
               srcId.begin() + 1, srcId.end(), 0LL,
-              [&](long long acc, size_t idx) { return acc + (*it)[idx]; });
+              [&](int64_t acc, size_t idx) { return acc + (*it)[idx]; });
           value -= sub_sum;
           (*it)[dstId] = static_cast<int>(value);
           ++counter;
